Move onboarding files
