+++ webapp/src/App.tsx
@@ -31,14 +31,9 @@ import Funding from "./pages/Funding";
 import FundingAddCard from "./pages/FundingAddCard";
 import FundingAddFunds from "./pages/FundingAddFunds";
 import FundingLinkBankAccount from "./pages/FundingLinkBankAccount";
-import Home from "./pages/Home";
 import LedgersOverview from "./pages/LedgersOverview";
 import MarkdownContent from "./pages/MarkdownContent";
 import Mobility from "./pages/Mobility";
-import Onboarding from "./pages/Onboarding";
-import OnboardingFinish from "./pages/OnboardingFinish";
-import OnboardingSignup from "./pages/OnboardingSignup";
-import OneTimePassword from "./pages/OneTimePassword";
 import OrderHistoryDetail from "./pages/OrderHistoryDetail";
 import OrderHistoryList from "./pages/OrderHistoryList";
 import PartnerSignup from "./pages/PartnerSignup";
@@ -48,12 +43,17 @@ import PrivacyPolicy from "./pages/PrivacyPolicy";
 import PrivateAccountDetail from "./pages/PrivateAccountDetail";
 import PrivateAccountsList from "./pages/PrivateAccountsList";
 import RegainAccountAccess from "./pages/RegainAccountAccess";
-import Start from "./pages/Start";
 import Styleguide from "./pages/Styleguide";
 import TripDetail from "./pages/TripDetail";
 import Trips from "./pages/Trips";
 import UnclaimedOrderList from "./pages/UnclaimedOrderList";
 import Utilities from "./pages/Utilities";
+import Home from "./pages/onboarding/Home";
+import Onboarding from "./pages/onboarding/Onboarding";
+import OnboardingFinish from "./pages/onboarding/OnboardingFinish";
+import OnboardingSignup from "./pages/onboarding/OnboardingSignup";
+import OneTimePassword from "./pages/onboarding/OneTimePassword";
+import Start from "./pages/onboarding/Start";
 import BackendGlobalsProvider from "./state/BackendGlobalsProvider";
 import GlobalViewStateProvider from "./state/GlobalViewStateProvider";
 import OfferingProvider from "./state/OfferingProvider";
