+++ webapp/src/pages/Styleguide.tsx
@@ -16,6 +16,7 @@ import Checklist from "../ui/Checklist";
 import ChecklistItem from "../ui/ChecklistItem";
 import Chip from "../ui/Chip";
 import Container from "../ui/Container";
+import DefinitionTable from "../ui/DefinitionTable.tsx";
 import { Dialog } from "../ui/Dialog";
 import DialogHeader from "../ui/DialogHeader";
 import Form from "../ui/Form.tsx";
@@ -33,7 +34,6 @@ import Switch from "../ui/Switch";
 import SwitchRow from "../ui/SwitchRow";
 import TextInput from "../ui/TextInput";
 import Tile from "../ui/Tile";
-import DefinitionTable from "./DefinitionTable.tsx";
 import BuildingStorefrontIcon from "@heroicons/react/24/outline/BuildingStorefrontIcon";
 import HomeIcon from "@heroicons/react/24/outline/HomeIcon";
 import ShoppingCartIcon from "@heroicons/react/24/outline/ShoppingCartIcon";
