+++ webapp/src/pages/onboarding/OnboardingSignup.tsx
@@ -1,18 +1,18 @@
-import api from "../api";
-import FormControlGroup from "../components/FormControlGroup";
-import OrganizationInputDropdown from "../components/OrganizationInputDropdown";
-import PageHeading from "../components/PageHeading";
-import { t } from "../localization";
-import keepDigits from "../modules/keepDigits";
-import useAsyncFetch from "../state/useAsyncFetch";
-import { extractErrorCode } from "../state/useError";
-import useUser from "../state/useUser";
-import Col from "../ui/Col";
-import Form from "../ui/Form";
-import FormButtons from "../ui/FormButtons";
-import FormError from "../ui/FormError";
-import FormSelect from "../ui/FormSelect";
-import Row from "../ui/Row";
+import api from "../../api";
+import FormControlGroup from "../../components/FormControlGroup";
+import OrganizationInputDropdown from "../../components/OrganizationInputDropdown";
+import PageHeading from "../../components/PageHeading";
+import { t } from "../../localization";
+import keepDigits from "../../modules/keepDigits";
+import useAsyncFetch from "../../state/useAsyncFetch";
+import { extractErrorCode } from "../../state/useError";
+import useUser from "../../state/useUser";
+import Col from "../../ui/Col";
+import Form from "../../ui/Form";
+import FormButtons from "../../ui/FormButtons";
+import FormError from "../../ui/FormError";
+import FormSelect from "../../ui/FormSelect";
+import Row from "../../ui/Row";
 import React from "react";
 import { useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
