+++ webapp/src/pages/onboarding/OneTimePassword.tsx
@@ -1,20 +1,20 @@
-import api from "../api";
-import FormSuccess from "../components/FormSuccess";
-import { t } from "../localization";
-import { dayjs } from "../modules/dayConfig";
-import { maskPhoneNumber } from "../modules/maskPhoneNumber";
-import { extractLocalizedError, useError } from "../state/useError";
-import useLoginRedirectLink from "../state/useLoginRedirectLink";
-import useUser from "../state/useUser";
-import BackButton from "../ui/BackButton.tsx";
-import BreadcrumbBack from "../ui/BreadcrumbBack.tsx";
-import Button from "../ui/Button";
-import ButtonGroup from "../ui/ButtonGroup.tsx";
-import ContinueButton from "../ui/ContinueButton.tsx";
-import Form from "../ui/Form";
-import FormError from "../ui/FormError";
-import Page from "../ui/Page.tsx";
-import Stack from "../ui/Stack.tsx";
+import api from "../../api";
+import FormSuccess from "../../components/FormSuccess";
+import { t } from "../../localization";
+import { dayjs } from "../../modules/dayConfig";
+import { maskPhoneNumber } from "../../modules/maskPhoneNumber";
+import { extractLocalizedError, useError } from "../../state/useError";
+import useLoginRedirectLink from "../../state/useLoginRedirectLink";
+import useUser from "../../state/useUser";
+import BackButton from "../../ui/BackButton.tsx";
+import BreadcrumbBack from "../../ui/BreadcrumbBack.tsx";
+import Button from "../../ui/Button";
+import ButtonGroup from "../../ui/ButtonGroup.tsx";
+import ContinueButton from "../../ui/ContinueButton.tsx";
+import Form from "../../ui/Form";
+import FormError from "../../ui/FormError";
+import Page from "../../ui/Page.tsx";
+import Stack from "../../ui/Stack.tsx";
 import "./OneTimePassword.css";
 import React from "react";
 import { useNavigate, useLocation } from "react-router-dom";
