+++ webapp/src/pages/onboarding/Start.tsx
@@ -1,19 +1,19 @@
-import api from "../api";
-import SignupAgreement from "../components/SignupAgreement";
-import { MdLink } from "../components/SumaMarkdown";
-import { t } from "../localization";
-import useI18n from "../localization/useI18n";
-import { dayjs } from "../modules/dayConfig";
-import { Logger } from "../modules/logger";
-import { extractErrorCode, extractLocalizedError, useError } from "../state/useError";
-import useToggle from "../state/useToggle";
-import BreadcrumbBack from "../ui/BreadcrumbBack.tsx";
-import ButtonGroup from "../ui/ButtonGroup.tsx";
-import ContinueButton from "../ui/ContinueButton.tsx";
-import Form from "../ui/Form";
-import FormError from "../ui/FormError";
-import Page from "../ui/Page.tsx";
-import PhoneInput from "../ui/PhoneInput";
+import api from "../../api";
+import SignupAgreement from "../../components/SignupAgreement";
+import { MdLink } from "../../components/SumaMarkdown";
+import { t } from "../../localization";
+import useI18n from "../../localization/useI18n";
+import { dayjs } from "../../modules/dayConfig";
+import { Logger } from "../../modules/logger";
+import { extractErrorCode, extractLocalizedError, useError } from "../../state/useError";
+import useToggle from "../../state/useToggle";
+import BreadcrumbBack from "../../ui/BreadcrumbBack.tsx";
+import ButtonGroup from "../../ui/ButtonGroup.tsx";
+import ContinueButton from "../../ui/ContinueButton.tsx";
+import Form from "../../ui/Form";
+import FormError from "../../ui/FormError";
+import Page from "../../ui/Page.tsx";
+import PhoneInput from "../../ui/PhoneInput";
 import React from "react";
 import { useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
