+++ webapp/src/ui/DefinitionTable.css
